# Backport upstream improvements to pg_dump