--- a/src/bin/pg_dump/common.c
+++ b/src/bin/pg_dump/common.c
@@ -626,6 +626,7 @@ AssignDumpId(DumpableObject *dobj)
 	dobj->dump = DUMP_COMPONENT_ALL;	/* default assumption */
 	dobj->dump_contains = DUMP_COMPONENT_ALL;	/* default assumption */
 	dobj->ext_member = false;	/* default assumption */
+	dobj->depends_on_ext = false;	/* default assumption */
 	dobj->dependencies = NULL;
 	dobj->nDeps = 0;
 	dobj->allocDeps = 0;