The purpose of this PR is to backport a series of refactors and performance improvements made to pg_dump from upstream REL_12_STABLE through REL_15_BETA1.

Before:
```
t/001_basic.pl ......... ok
t/002_pg_dump.pl ....... ok
t/010_dump_connstr.pl .. ok
All tests successful.
Files=3, Tests=6093, 41 wallclock secs ( 0.19 usr  0.02 sys +  5.46 cusr  0.99 csys =  6.66 CPU)
Result: PASS
pg_dump -s regression > /dev/null  0.59s user 0.12s system 2% cpu 33.560 total
```

After:
```
t/001_basic.pl ......... ok
t/002_pg_dump.pl ....... ok
t/010_dump_connstr.pl .. ok
All tests successful.
Files=3, Tests=6248, 27 wallclock secs ( 0.19 usr  0.02 sys +  5.47 cusr  0.95 csys =  6.63 CPU)
Result: PASS
pg_dump -s regression > /dev/null  0.44s user 0.07s system 6% cpu 8.440 total
```

Below is a list of the backported upstream commits. Please refer to individual commit messages for details and resolved conflicts.


https://github.com/postgres/postgres/commit/d5e8930 pg_dump: minor performance improvements from eliminating sub-SELECTs.
https://github.com/postgres/postgres/commit/e2c52be pg_dump: Refactor getIndexes()
https://github.com/postgres/postgres/commit/be85727 Use PREPARE/EXECUTE for repetitive per-object queries in pg_dump.
https://github.com/postgres/postgres/commit/9895961 Avoid per-object queries in performance-critical paths in pg_dump.
https://github.com/postgres/postgres/commit/0c9d844 Rethink pg_dump's handling of object ACLs.
https://github.com/postgres/postgres/commit/5209c0b Refactor pg_dump's tracking of object components to be dumped.
https://github.com/postgres/postgres/commit/a22d6a2 pg_dump: Add missing relkind case
https://github.com/postgres/postgres/commit/70bef49 Fix minor memory leaks in pg_dump.
https://github.com/postgres/postgres/commit/92316a4 In pg_dump, use simplehash.h to look up dumpable objects by OID.
https://github.com/postgres/postgres/commit/2acc84c pg_dump: fix mis-dumping of non-global default privileges.
https://github.com/postgres/postgres/commit/4438eb4 pg_dump: Reorganize getTables()
https://github.com/postgres/postgres/commit/6b96673 In pg_dump, avoid doing per-table queries for RLS policies.
https://github.com/postgres/postgres/commit/6c450a8 Cache the results of format_type() queries in pg_dump.
https://github.com/postgres/postgres/commit/d782d59 Avoid invoking PQfnumber in loop constructs
https://github.com/postgres/postgres/commit/6499008 pg_dump: Add const decorations
https://github.com/postgres/postgres/commit/0bf8364 pg_dump: Fix dumping of inherited generated columns
https://github.com/postgres/postgres/commit/8e396a7 pg_dump: label PUBLICATION TABLE ArchiveEntries with an owner.
https://github.com/postgres/postgres/commit/9eabfe300a22ad3d776dc293265e15379790bd9a pg_dump: label INDEX ATTACH ArchiveEntries with an owner.
https://github.com/postgres/postgres/commit/d5ab79d815783fe60062cefc423b54e82fbb92ff Make pg_dump's table of object-type priorities more maintainable.
https://github.com/postgres/postgres/commit/04732962462ba99cf8f8bcf6ac83932867cc96a8 pg_dump: Reorganize dumpBaseType()
https://github.com/postgres/postgres/commit/8354e7b27ebec8bf78088e22b5b9c1caa46db8e1 Remove unused parameters
https://github.com/postgres/postgres/commit/5423853feebd30772b7ff9b306885dcb02b79e76 Avoid retrieval of CHECK constraints and DEFAULT exprs in data-only dump
https://github.com/postgres/postgres/commit/ed2c7f65bd9f15f8f7cd21ad61602f983b1e72e9 pg_dump: Reorganize dumpFunc() and dumpAgg()
https://github.com/postgres/postgres/commit/5fea14f4b2c5ca4e4d9def92b17e50fcfbaf3468 Avoid trying to restore table ACLs and per-column ACLs in parallel.
https://github.com/postgres/postgres/commit/991c444e7ad3a1ff7bcffff4121e6ff67c4783fc pg_dump: Further reorganize getTableAttrs()
https://github.com/postgres/postgres/commit/d9fa17aa7c34dea66ce64da6fb4c643e75ba452c pg_dump: Remove dead code
https://github.com/postgres/postgres/commit/b08dee24a5574ba6657d238b91c10f1f8e2608f8 Add pg_dump support for ALTER obj DEPENDS ON EXTENSION
https://github.com/postgres/postgres/commit/f31364676dfd9e7a2b712ea0da7b4ab6b7d28f5e Teach pg_dump to dump comments on RLS policy objects.
https://github.com/postgres/postgres/commit/8f4ee44bcdbcd437f860f2d048125a118335b088 Fix minor memory leak in pg_dump
https://github.com/postgres/postgres/commit/07b39083c2aca003c4b1f289d7dc2368b5e2297a Fix pg_dump's handling of dependencies for custom opclasses.
 postgres/postgres@9a4c0e3 Dump ALTER TABLE ... ATTACH PARTITION as a separate ArchiveEntry.
 postgres/postgres@d5ab79d Make pg_dump's table of object-type priorities more maintainable.
 postgres/postgres@9e5f1f2 Rethink recent fix for pg_dump's handling of extension config tables.
 postgres/postgres@3eb3d3e Collect attribute data on extension owned tables being dumped
 postgres/postgres@3dadcb2 pg_dump: fix dependencies on FKs to partitioned tables
 postgres/postgres@b304b2b Fix parallel restore of FKs to partitioned tables
 
In addition to the backports:
* Pre-GPDB5 version specific code has been removed.
* The GPDB specific TypeStorageOptions struct and supporting code has been removed in favor of adding
  pg_type_encoding.typoptions to the getTypes query. The usage of CatalogId as the hash index caused collisions
  with the TypeInfo indexes when TypeStorageOptions was a unique dumpable object, since the only unique aspect was the
  typstorage field. Now, when dumpType is called, we check the typstorage field and dump the ALTER TABLE...SET DEFAULT 
  ENCODING TOC entry if needed.
* The following unused flags have been removed:
  - pre-data-schema-only
  - post-data-schema-only